Added log out function, updated amount field for decimal and updates to refresh home page after inserting new records

+++ expense_tracker/expense_tracker_app/models.py
@@ -8,7 +8,7 @@ from django.db import models
 
 class ExpenseDataModel(models.Model):
     expense_category = models.CharField(max_length=100)
-    expense_amount = models.IntegerField()
+    expense_amount = models.DecimalField(max_digits=20, decimal_places=3)
     expense_date = models.DateTimeField()
     user = models.ForeignKey(settings.AUTH_USER_MODEL, on_delete=models.CASCADE, default=None)
